Refactor the project to align with Maven
The project, while valid, diverges from the module lessons and then
Maven as presented. With the junit5-jars provided, IDEs (like IntelliJ)
attempts to add them as a module within the project.

IntelliJ is intelligent enough to determine the source and test
directories, but including the test file, deviates from the presented
content. Deleting this allows the branch to start in a place ready for
the content.

Aliging with Maven is actually critical as module 2 later suggests this,
however, the project layout would create issues for Maven without
specific build overrides to the Super POM for src and test locations.

Moving the files to align with Maven convention sets up learners with a
familiar source code layout as well. Without requiring a pom.xml.